optimize foreign key for Gig model

+++ Gig/Gig/ViewModel/GigFormViewModel.cs
@@ -10,7 +10,7 @@ namespace Gig.ViewModels
 
         public string Date { get; set; }
 
-        public int Genre { get; set; }
+        public byte Genre { get; set; }
         public IEnumerable<Genre> Genres { get; set; }
     }
 }
